Ajout d'un message sur aucun topic créé

+++ view/forum/listTopics.php
@@ -15,28 +15,36 @@ $topics = $result["data"]['topics'];
     </div>
 
     <?php
-    foreach($topics as $topic ){
+    if(!empty($topics)){
+        
+        foreach($topics as $topic ){
+    
+            ?>
+            <div class="listContainer">
+                <div class="leftTopicContainer">
+                    <div class="TitleAndCategory">
+                        <h2><a href="index.php?ctrl=forum&action=listPosts&id=<?=$topic->getId()?>"><?=$topic->getTitle()?></a></h2> 
+                        <p>&nbsp- <?=$topic->getCategory()->getCategoryName() ?></p>
+                        
+                    </div>
+                    <div class="additionnalContent">
+                        <p >date de création : le <?=$topic->getCreationDate()->format("d-m-Y à H:i")?> - </p>
+                        <p><a href="index.php?ctrl=forum&action=index&deleteTopic&id=<?=$topic->getId()?>">Supprimer</a></p>
+                    </div>
+                </div>
+                <div class="rightTopicContainer">
+                    <p>par <span class="user"><?=$topic->getUser()->getPseudo()?></span></p>
+                    <p>Nombre de posts : <?=$topic->getNbPosts() ?></p>
+                </div>
+            </div>
+            <?php
+        } 
 
-        ?>
+    }else {?>
         <div class="listContainer">
-            <div class="leftTopicContainer">
-                <div class="TitleAndCategory">
-                    <h2><a href="index.php?ctrl=forum&action=listPosts&id=<?=$topic->getId()?>"><?=$topic->getTitle()?></a></h2> 
-                    <p>&nbsp- <?=$topic->getCategory()->getCategoryName() ?></p>
-                    
-                </div>
-                <div class="additionnalContent">
-                    <p >date de création : le <?=$topic->getCreationDate()->format("d-m-Y à H:i")?> - </p>
-                    <p><a href="index.php?ctrl=forum&action=index&deleteTopic&id=<?=$topic->getId()?>">Supprimer</a></p>
-                </div>
-            </div>
-            <div class="rightTopicContainer">
-                <p>par <span class="user"><?=$topic->getUser()->getPseudo()?></span></p>
-                <p>Nombre de posts : <?=$topic->getNbPosts() ?></p>
-            </div>
+            <p>aucun topic créé pour le moment...</p>
         </div>
-        <?php
-    } ?>
+    <?php } ?>
 </div>
